+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 * 2017-03-07:  3.4.1
   - LOD adjustments/fixes:
+    - use cfg.portal.mainPage instead of windowHref in some filters dealing with creating uri links
     - mkLinks filter: use getHref4uriWithSelfHostPrefix (instead of older mechanism `/#uri/...`)
   - fix #97 "create vocab/mapping dialog stays open when user navigates back in browser".
     Simply cancelUriEdit() upon $stateChangeStart